feat(event-handler): Add AppSync handler decorator
