+++ aws_lambda_powertools/utilities/event_handler/__init__.py
@@ -1,0 +1,3 @@
+from .appsync import AppSyncResolver
+
+__all__ = ["AppSyncResolver"]
